testqwen/dock2/requirements.txt
torch
torchvision
torchaudio
torchao
git+https://github.com/huggingface/transformers.git
diffusers
Pillow
byaldi
qwen_vl_utils
flash-attn
fastapi
uvicorn[standard]